Set correct minSize for QRCodes
