+++ visualization/app/codeCharta/services/3DExports/3DPreview/MeshModels/BackMeshModels/qrCodeMesh.ts
@@ -1,19 +1,17 @@
 import { CustomVisibilityMesh } from "../customVisibilityMesh"
 import { GeometryOptions } from "../../preview3DPrintMesh"
-import { CreateQRCodeGeometryStrategy } from "../../CreateGeometryStrategies/createQRCodeGeometryStrategy"
 import { BackPrintColorChangeStrategy } from "../../ColorChangeStrategies/backPrintColorChangeStrategy"
+import { BoxGeometry, BufferGeometry } from "three"
+import { BufferGeometryUtils } from "three/examples/jsm/utils/BufferGeometryUtils"
+import * as QRCode from "qrcode"
 
 export class QrCodeMesh extends CustomVisibilityMesh {
-    //TODO: make this implement GeneralSizeChangeMesh and update the errorCorrectionLevel - for this some 3D printing tests need to be made before
-    private readonly createQRCodeStrategy: CreateQRCodeGeometryStrategy
-
     constructor(name: string) {
-        super(name, new BackPrintColorChangeStrategy(), 280, false)
-        this.createQRCodeStrategy = new CreateQRCodeGeometryStrategy()
+        super(name, new BackPrintColorChangeStrategy(), 0, false)
     }
 
     async init(geometryOptions: GeometryOptions): Promise<QrCodeMesh> {
-        this.geometry = await this.createQRCodeStrategy.create(geometryOptions)
+        this.geometry = await this.create(geometryOptions)
 
         const xPosition = 0.45
         const yPosition = 0.45
@@ -26,6 +24,41 @@ export class QrCodeMesh extends CustomVisibilityMesh {
     }
 
     async changeText(geometryOptions: GeometryOptions): Promise<void> {
-        this.geometry = await this.createQRCodeStrategy.create(geometryOptions)
+        this.geometry = await this.create(geometryOptions)
+    }
+
+    async create(geometryOptions: GeometryOptions): Promise<BufferGeometry> {
+        if (!geometryOptions.qrCodeText || geometryOptions.qrCodeText.length === 0) {
+            return new BufferGeometry()
+        }
+
+        const canvas = document.createElement("canvas")
+        await QRCode.toCanvas(canvas, geometryOptions.qrCodeText, { errorCorrectionLevel: "H" }) //high error correction works (counter intuitively) way better than low error correction
+
+        const context = canvas.getContext("2d")
+        const imageData = context.getImageData(0, 0, canvas.width, canvas.height)
+
+        const qrCodeGeometries: BufferGeometry[] = []
+        const pixelSize = 0.6 / imageData.width
+        // Loop over each pixel in the image
+        for (let y = 0; y < imageData.height; y += 4) {
+            for (let x = 0; x < imageData.width; x += 4) {
+                const index = (y * imageData.width + x) * 4
+                if (imageData.data[index] !== 0) {
+                    const geometry = new BoxGeometry(pixelSize, pixelSize, geometryOptions.printHeight)
+                    geometry.translate((-x / 4) * pixelSize, (-y / 4) * pixelSize, 0)
+                    qrCodeGeometries.push(geometry)
+                }
+            }
+        }
+
+        this.updateMinWidth(pixelSize)
+
+        return BufferGeometryUtils.mergeBufferGeometries(qrCodeGeometries)
+    }
+
+    private updateMinWidth(pixelSize: number) {
+        const minWidthOfOneQRCodePixel = 0.6 //mm I tested this value on the prusaXL with 0.4mm nozzle
+        this.minWidth = minWidthOfOneQRCodePixel / pixelSize
     }
 }
